Removing pointer-events from BackToTop wrapper
The back-to-top wrapper element is hidden but blocks hover events from other elements that may overlap in the same position.
NOTE that this does not affect any component in the current design. I happened to notice this on a different fork where it was causing an issue and therefore I suggest the change but feel free to ignore this if you think is not relevant enough or could cause problems with the back to top component.